requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
scikit-learn
pandas
joblib
transformers
torch
